+++ server/src/main/java/com/photoday/photoday/excpetion/ExceptionCode.java
@@ -11,7 +11,8 @@ public enum ExceptionCode {
     CANNOT_FOLLOW_MYSELF(HttpStatus.UNAUTHORIZED, "본인을 팔로우 할 수 없습니다"),
     ALREADY_REPORTED(HttpStatus.CONFLICT, "이미 신고한 게시물입니다."),
     IMAGE_NOT_FOUND(HttpStatus.NOT_FOUND, "이미지가 없습니다."),
-    BANNED_USER(HttpStatus.FORBIDDEN, "정지된 회원입니다.");
+    ACCOUNT_SUSPENDED(HttpStatus.FORBIDDEN, "부적절한 이용으로 계정이 일시 정지되었습니다."),
+    PASSWORD_NOT_MATCH(HttpStatus.UNAUTHORIZED, "비밀번호가 일치하지 않습니다.");
 
     @Getter
     private HttpStatus httpStatus;
